requirements.txt
openai==1.82.0
gspread==6.2.0
google-auth==2.38.0
apscheduler==3.11.0
pydantic==2.11.1
pydantic-settings==2.8.1
requests==2.32.3
python-dotenv==1.1.0
